Fix LEDs
